requirements.txt
APScheduler==3.7.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Deprecated==1.2.12
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
psycopg2==2.8.6
PyGithub==1.54.1
PyJWT==1.7.1
pyparsing==2.4.7
python-dotenv==0.17.0
pytz==2021.1
requests==2.25.1
six==1.15.0
spotipy==2.17.1
SQLAlchemy==1.4.6
sqlalchemy-redshift==0.8.2
tzlocal==2.1
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
